+++ tests/GaussianEliminationTest.cc
@@ -5,7 +5,7 @@
 namespace caramel::tests {
 
 TEST(GaussianEliminationTest, TestSimple) {
-  std::vector<std::vector<uint32_t>> matrix = {{0, 1, 2}, {1, 2}, {0, 2}};
+  std::vector<std::vector<uint64_t>> matrix = {{0, 1, 2}, {1, 2}, {0, 2}};
   std::vector<uint32_t> constants = {1, 0, 1};
   std::string solution_str = "100";
 
@@ -20,7 +20,7 @@ TEST(GaussianEliminationTest, TestSimple) {
 }
 
 TEST(GaussianEliminationTest, TestWithSwaps) {
-  std::vector<std::vector<uint32_t>> matrix = {{1, 34}, {0, 1, 34}, {0, 34}};
+  std::vector<std::vector<uint64_t>> matrix = {{1, 34}, {0, 1, 34}, {0, 34}};
   std::vector<uint32_t> constants = {0, 1, 1};
   std::string solution_str = "1";
   for (uint32_t i = 0; i < 34; i++) {
@@ -38,7 +38,7 @@ TEST(GaussianEliminationTest, TestWithSwaps) {
 }
 
 TEST(GaussianEliminationTest, TestEmpty) {
-  std::vector<std::vector<uint32_t>> matrix = {{1, 34}, {0, 1, 34}, {0, 34}};
+  std::vector<std::vector<uint64_t>> matrix = {{1, 34}, {0, 1, 34}, {0, 34}};
   std::vector<uint32_t> constants = {0, 1, 1};
   std::string solution_str;
   for (uint32_t i = 0; i < 35; i++) {
@@ -56,7 +56,7 @@ TEST(GaussianEliminationTest, TestEmpty) {
 }
 
 TEST(GaussianEliminationTest, TestUnsolvable) {
-  std::vector<std::vector<uint32_t>> matrix = {{0, 1}, {0, 1}, {0, 2}};
+  std::vector<std::vector<uint64_t>> matrix = {{0, 1}, {0, 1}, {0, 2}};
   std::vector<uint32_t> constants = {0, 1, 1};
 
   auto system = DenseSystem::make(3);
@@ -71,7 +71,7 @@ TEST(GaussianEliminationTest, TestUnsolvable) {
 }
 
 TEST(GaussianEliminationTest, TestSettingMultipleSolutionBits) {
-  std::vector<std::vector<uint32_t>> matrix = {{0, 1}, {1, 2}, {2, 3}, {3, 4}};
+  std::vector<std::vector<uint64_t>> matrix = {{0, 1}, {1, 2}, {2, 3}, {3, 4}};
   std::vector<uint32_t> constants = {1, 1, 1, 1};
   std::string solution_str = "01010";
